requirements.txt
pandas
gradio
transformers
sentence-transformers
google-generativeai
scikit-learn